### 2019-06-16 #1507 @arfon
Hi @ramaniitrgoyal92, thanks for your submission to JOSS. Please rename your paper to `paper.md`
