### 2019-06-16 #1507 @ramaniitrgoyal92
> **Submitting author:** @ramaniitrgoyal92 ([Raman Goyal](http://orcid.org/0000-0002-8128-3051))
> **Repository:** https://github.com/ramaniitrgoyal92/Tensegrity_Engineering_Analysis_Master-TEAM-
> **Version:** v2.1
> **Editor:** Pending
> **Reviewer:** Pending
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @ramaniitrgoyal92. **Currently, there isn't an JOSS editor assigned** to your paper.
>
> @ramaniitrgoyal92 if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
>
> ```
> @whedon commands
> `

The suggested reviewers:

@apsabelhaus
@moorepants
@ctdegroot
@vaishnavtv
@datseris
