### 2019-06-16 #1507 @apsabelhaus
For the editor, once they're assigned:
If chosen as a reviewer, I'd be happy to review this work.

However, I'd like to quickly check about JOSS' Conflict of Interest policy (https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), since I'm still new to open review (not blind). The submitting author @ramaniitrgoyal92 and I are professional acquaintances who have discussed research before and compared our past publications, but have not collaborated on funded work or authorship, and are not members of the same institution. Normally this wouldn't be a problem for single-blind review, but does it pose a challenge for open review? Either way, I believe I can provide an "impartial assessment," but would like to disclose our relationship before agreeing to be a reviewer.
