### 2019-06-17 #1507 @Kevin-Mattheus-Moerman
@vaishnavtv @ctdegroot since you were involved with https://github.com/openjournals/joss-reviews/issues/1042, would you be interested in helping with this review?
