### 2019-06-17 #1507 @Kevin-Mattheus-Moerman
@ramaniitrgoyal92 Please rename your paper to `paper.md`, this way `@whedon` will find it okay to render the PDF. Thanks
