### 2019-06-17 c0721cd
Update paper.md
