### 2019-06-17 a50a1e1
Update paper.md
