+++ JOSS_Paper/paper.md
@@ -1,3 +1,5 @@
+---
+
 title: 'TEAM: Tensegrity Engineering Analysis Master'
 
 tags:
