### 2019-06-17 6e44596
Update paper.md
