+++ JOSS_Paper/paper.md
@@ -24,7 +24,7 @@ authors:
     affiliation: 1
 
   - name: Robert E. Skelton
-    orcid: 
+    orcid: 0000-0001-6503-9115
     affiliation: 1
     
 affiliations:
