### 2019-07-03 #1507 @Kevin-Mattheus-Moerman
@Jfriesen222 would you be interested in reviewing this work on tensegrity modelling for [JOSS](http://joss.theoj.org)? The review process will focus on [the software](https://github.com/ramaniitrgoyal92/Tensegrity_Engineering_Analysis_Master-TEAM-) and [this short paper](https://github.com/openjournals/joss-papers/blob/joss.01507/joss.01507/10.21105.joss.01507.pdf).
