### 2019-07-27 #1507 @apsabelhaus
@ramaniitrgoyal92 while we're waiting for the official review to open, I added a first issue in the repo about organization and documentation, so there's something to start on! Also, the paper is missing references [???], probably a bug in the latex.
