### 2019-07-31 #1507 @Kevin-Mattheus-Moerman
@Jfriesen222 would you be interested in reviewing this work on tensegrity modelling for JOSS? The review process will focus on the software and this short paper.
