### 2019-07-31 #1613 @whedon
[REVIEW]: TEAM: Tensegrity Engineering Analysis Master
**Submitting author:** @ramaniitrgoyal92 (<a href="http://orcid.org/0000-0002-8128-3051">Raman Goyal</a>)
**Repository:** <a href="https://github.com/ramaniitrgoyal92/Tensegrity_Engineering_Analysis_Master-TEAM-" target ="_blank">https://github.com/ramaniitrgoyal92/Tensegrity_Engineering_Analysis_Master-TEAM-</a>
**Version:** v2.1
