+++ setup.m
@@ -1,4 +1,4 @@
-%% ^_^ Welcome to Tensegrity Engineering Analysis Master(TEAM) software! ^_^ %%
+%% ^_^ Welcome to Modeling of Tensegrity Structrues(MOTES) software! ^_^ %%
 
 %% SETUP file to be run only the first time
 
@@ -6,15 +6,17 @@
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
 
-%% Add all necessary functions to MATLAB path
-clear all; clc; close all
-addpath 'Main' 'Test_Examples' 'Tensegrity_Examples' 'User_Guide' 'Videos' 'JOSS_Paper';
-% Add subfolders of 'Main'
-cd Main;
-addpath 'Tensegrity_Statics_Main' 'Tensegrity_Dynamics_Main';
-cd ..;
-%% Open the User_Guide
-cd User_Guide;
-open('User_Guide_Tensegrity_Engineering_Analysis_Master_(TEAM).pdf');
-cd ..;
-disp('Welcome! Please follow the step-by-step instructions from the User Guide.');
+%% Add all necessary functions to MATLAB path%% set up the workspace
+clear all;
+close all;
+clc;
+% add the function libraries
+addpath( genpath('Function_Library') );
+% add the Software Verification and Examples
+addpath( genpath('Software_Verification_and_Examples') );
+% add User Guide
+addpath( genpath('User_Guide') );
+% add Videos folder
+addpath( genpath('Videos') );
+% add JOSS Paper
+addpath( genpath('JossPaper') );
