### 2019-08-06 4245641
Delete PlotCylinderObject.m
