### 2019-08-06 0a54b0b
Delete bar_length_correction_Class1.m
