### 2019-08-06 e874ec1
Delete bar_length_correction_ClassK.m
