### 2019-08-06 b0b775b
Delete extractSignals.m
