### 2019-08-06 bad0ebb
Delete ode4.m
