### 2019-08-06 443027e
Delete tenseg_animation.m
