### 2019-08-06 895aacd
Delete tenseg_class_k_convert.m
