### 2019-08-06 e6193a8
Delete tenseg_defaults.m
