### 2019-08-06 2e219dc
Delete tenseg_dyn_c1open_fnc.m
