+++ Main/Tensegrity_Dynamics_Main/tenseg_dyn_c1open_fnc.m
@@ -1,221 +1,0 @@
-function [ Xd, logging] = tenseg_dyn_c1open_fnc( t,X,INPUTS )
-% /* This Source Code Form is subject to the terms of the Mozilla Public
-% * License, v. 2.0. If a copy of the MPL was not distributed with this
-% * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
-% [ Xd, logging] = TENSEG_DYN_C1OPEN_FNC( t,X,INPUTS ) is used in computing
-% the dynamical behavior of an open-loop class 1 tensegrity structure.
-% Essentially, X_dot = [Nd; Ndd] is computed from X = [N, Nd], and t is
-% included for use with ode4, a fixed time-step RK4 integrator.
-% Structure-specific definitions are provided through INPUTS, and
-% functionality is included that allows for the logging of internally
-% computed variables throughout a simulation, such as bar length and string
-% force densities. At a high level, this internal logging works by storing
-% the internal variables in a persistent internal variable. If the function
-% is then called without any inputs, the logged internal values are
-% returned.
-%
-% Inputs:
-%	t: current time (for use with ode4)
-%	X: current N and N_dot values
-%		stacked column-wise into a vector: X = [N; Nd];
-%	INPUTS: structure containing tensegrity structure-specific information
-%		below fields required from TENSEG_SIM_CLASS1OPEN.m
-%		[].constants
-%		[].sim
-%		[].forces
-%		[].tenseg_struct
-%
-% Outputs:
-%	Xd: N_dot and N_dotdot values based on current N and N_dot. Again,
-%		stacked column-wise into a vector Xd = [Nd; Ndd]
-%	logging: used for outputting internally logged variables
-
-
-% Used to log internal variables. Remain in workspace through multiple
-% function calls
-persistent i logging_temp
-
-
-% Prep for logging internal variables -------------------------------------
-% Labels for internally-logged data
-bar_len_log = 1;
-gamma_log = 2;
-
-% Check if this is the start of a simulation and initialize logging arrays
-if nargout == 1 && (isempty(i) || i > 2) && t==0
-	i = 1;
-	
-	% Read in some inputs to initialize size of internal logging arrays
-	constants = INPUTS.constants;
-	sim = INPUTS.sim;
-	
-	% Initialize persistent logging arrays
-	logging_temp{bar_len_log} = zeros(constants.beta,numel(sim.tspan));
-	logging_temp{gamma_log} = zeros(constants.alpha,numel(sim.tspan));
-end
-
-
-% If simulation (one argument output), read inputs and compute Xd
-if nargout == 1
-	
-	%disp(t)
-	
-	idx = 0.25 * (i + 3); % used for checking if we're on a reported time-step
-	ind_use = floor((i-2)/4)+1; % index to use when reading in time-varying info
-	if ind_use==0
-		ind_use=1;
-	end
-	dt = INPUTS.sim.dt;
-	
-	% Read inputs
-	forces = INPUTS.forces;
-	tenseg_struct = INPUTS.tenseg_struct;
-	constants = INPUTS.constants;
-
-	Wgiven = forces.W;
-	C_b = tenseg_struct.C_b;
[…]
