### 2019-08-06 e2764a8
Delete tenseg_equilibrium.m
