### 2019-08-06 128b9b4
Delete tenseg_dyn_ckopen_fnc.m
