### 2019-08-06 ec37ed7
Delete tenseg_ind2C.m
