### 2019-08-06 d7f2db7
Delete tenseg_merge.m
