### 2019-08-06 6e9768e
Delete tenseg_struct_gen.m
