+++ Main/Tensegrity_Dynamics_Main/tenseg_struct_gen.m
@@ -1,230 +1,0 @@
-function [full_struct] = tenseg_struct_gen(input_struct)
-
-% /* This Source Code Form is subject to the terms of the Mozilla Public
-% * License, v. 2.0. If a copy of the MPL was not distributed with this
-% * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
-% [tenseg_struct] = TENSEG_STRUCT_GEN(tenseg_struct) fully populates all
-% fields of a tensegrity simulation task data structure given an initial
-% data structure containing, at a minimum, the required fields. Default
-% values are loaded and added as fields where necessary.
-%
-% Inputs:
-%	input_struct: tensegrity simulation task specification data structure
-%		REQUIRED FIELDS:
-%			[].N: initial node positions
-%			[].C_b: bar connectivity
-%			[].C_s: string connectivity
-%			[].s_0: resting string lengths
-%		OPTIONAL FIELDS (default values specified in 'tenseg_defaults'):
-%			[].P: constraint matrix
-%			[].D: constraint matrix
-%			[].Nd0: initial node velocities
-%			[].m: bar masses
-%			[].ms: string node point masses
-%			[].mgyro: gyro wheel masses
-%			[].gyro_r: gyro wheel radii
-%			[].gyro_h: gyro wheel heights
-%			[].tf: simulation time duration
-%			[].dt: simulation time step
-%			[].W: external node forces
-%			[].k: string stiffness coefficients
-%			[].gyro_omega: gyro angular rates
-%
-% Outputs:
-%	full_struct: summary of internal variables used
-%		[].input: input variables given (see input field descriptions)
-%		[].sys_def: tensegrity system definition variables
-%			[].[].C_sb: string connections to bars
-%			[].[].C_ss: string connections to string nodes
-%			[].[].C_bb: non-zero columns of C_b (old C_b matrix)
-%			[].[].C_nb: satisfies R_b = N*C_nb^T*C_r'
-%			[].[].C_s: overall string connectivity
-%			[].[].C_b: overall bar connectivity
-%			[].[].m: bar masses
-%			[].[].ms: string node point masses
-%			[].[].mgyro: gyro wheel masses
-%			[].[].gyro_r: gyro wheel radii
-%			[].[].gyro_h: gyro wheel heights
-%			[].[].C_r: Bar center of mass 'connectivity'
-%			[].[].len_hat: initial bar lengths
-%			[].[].Jt_hat: transaxial bar+gyro moments of inertia (/len?)
-%			[].[].Ja_hat: axial bar+gyro moments of inertia
-%		[].constants: various constants throughout the simulation
-%			[].[].n: number of nodes
-%			[].[].beta: number of bars
-%			[].[].alpha: number of string members
-%			[].[].sigma: number of string point mass nodes
-%			[].[].M: mass matrix used in dynamics
-%			[].[].Minv: inverse of M
-%		[].forces: force-related variables
-%			[].[].W: external node forces
-%			[].[].k: string stiffness coefficients
-%			[].[].s_0: resting string lengths
-%			[].[].gyro_omega: gyro angular rates
-%		[].ICs: initial conditions
-%			[].[].N0: initial node positions
-%			[].[].Nd0: initial node velocities
-%		[].sim: integration specification variables
-%			[].[].dt: simulation time step
-%			[].[].tf: simulation time duration
-
-
-% Save original input fields
-full_struct.input = input_struct;
-
-
[…]
