### 2019-08-06 961da6d
Delete tenseg_plot.m
