+++ Main/Tensegrity_Dynamics_Main/tenseg_plot.m
@@ -1,245 +1,0 @@
-function [fig_out] = tenseg_plot( N,C_b,C_s,fig_handle,highlight_nodes,view_vec, PlotTitle, R3Ddata)
-% /* This Source Code Form is subject to the terms of the Mozilla Public
-% * License, v. 2.0. If a copy of the MPL was not distributed with this
-% * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
-% [fig_out] = TENSEG_PLOT( N,C_b,C_s,fig_handle,highlight_nodes,view_vec )
-% creates a rough visualization figure for a given tensegrity structure
-%
-% Inputs:
-%	N: node matrix (3 x n array for n nodes)
-%	C_b (optional): bar connectivity matrix (beta x n array for beta bars)
-%	C_s: string connectivity matrix (alpha x n array for alpha strings)
-%	fig_handle (optional): figure in which to plot
-%	highlight_nodes (optional): node(s) to highlight when plotting (vector
-%		containing node numbers)
-%	view_vec (optional): plotting view direction (see view())
-%   PlotTitle (optional): Title of the plot, default: ''
-%   R3Ddata (optional): Structure with the radius of objects for 3D plots
-%        [].Bradius: Radius of bars [# bars x 1]
-%        [].Sradius: Radius of strings [# strings x 1]
-%        [].Nradius: Radius of node spheres [# nodes x 1]
-%
-% Outputs:
-%	fig_out: figure handle in which plot was made
-%
-% Example:
-%	tenseg_plot(N,C_b,C_b)
-
-% Handle optional arguments
-
-%% Object size options (for line plots)
-BarWidth = 2; % Width of bar lines
-StringWidth = 1; % Width of string lines
-NodeSize = 8; % Size of node marker
-
-%% Labeling options
-% Write labels? (1: show, 0: suppress)
-LabelNodes = 0;
-LabelStrings = 0;
-LabelBars = 0;
-
-FontBars = 15; % Font of bar labels
-FontStrings = 10; % Font of string labels
-FontNodes = 18; % Font of node labels
-
-FractionDistance = 0.005; % Distance between object and label (relative to overall axis size)
-
-%% 3D plot options
-nsurfpatches = 6; % Number of surface patches in 3D plots
-BarSurfColor = [0.2, 0.2, 0.2];
-StringSurfColor = [0.9, 0.1, 0.1];
-LightAmbientStrength = .7;% [0,1], 0.3 is Matlab's default
-
-%%
-switch nargin
-    case 5
-        view_vec = [];
-    case 4
-        view_vec = [];
-        highlight_nodes = [];
-    case 3
-        view_vec = [];
-        highlight_nodes = [];
-        fig_handle = [];
-end
-
-if nargin < 7 % Empty Plot Title
-    PlotTitle = '';
-end
-if nargin < 8 % No 3D object data
-    R3Ddata = [];
-end
-
-% Get min and max X,Y,Z coordinates contained in Nhist
-min_x = min(N(1,:,:)); max_x = max(N(1,:,:));
-min_y = min(N(2,:,:)); max_y = max(N(2,:,:));
-min_z = min(N(3,:,:)); max_z = max(N(3,:,:));
-
-%%
-if ~isempty(R3Ddata) % Extending axis to account for radii of objects
-    maxradius = 0;
-    if isfield(R3Ddata,'Bradius')
-        maxradius = max(maxradius,max(R3Ddata.Bradius));
-    end
-    if isfield(R3Ddata,'Sradius')
-        maxradius = max(maxradius,max(R3Ddata.Sradius));
-    end
[…]
