### 2019-08-06 c9c7306
Delete tenseg_plot_configurations.m
