### 2019-08-06 c29dc85
Delete tenseg_plot_node.m
