### 2019-08-06 b46dc46
Delete tenseg_plot_velocity.m
