### 2019-08-06 72afcce
Delete tenseg_reordering_topology.m
