+++ Main/Tensegrity_Dynamics_Main/tenseg_reordering_topology.m
@@ -1,397 +1,0 @@
-% /* This Source Code Form is subject to the terms of the Mozilla Public
-% * License, v. 2.0. If a copy of the MPL was not distributed with this
-% * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
-% Re-orders the node indeces if they are not ordered correctely.
-% This makes the bars connectivity matrix equal to [-I I 0] in the general
-% case
-% Also adds virtual nodes for any class-k and pinned node
-
-tenseg_struct_scrambled = tenseg_struct;
-ICs_scrambled = ICs;
-forces_scrambled = forces;
-
-C_b = field_to_var(tenseg_struct_scrambled,'C_b');
-C_s = field_to_var(tenseg_struct_scrambled,'C_s');
-N0 = field_to_var(ICs_scrambled,'N0');
-Nd0 = field_to_var(ICs_scrambled,'Nd0');
-W = field_to_var(forces_scrambled,'W');
-
-
-if exist('tenseg_struct_scrambled.pinned_nodes','var'),
-%     pinned_nodes_scr = field_to_var(tenseg_struct_scrambled,'pinned_nodes');
-pinned_nodes_scr = tenseg_struct_scrambled.pinned_nodes_scr;
-end
-
-nnodes_or = size(N0,2);
-nbars = size(C_b,1);
-nstrings = size(C_s,1);
-
-sconn_or = zeros(nstrings,2);
-
-for i = 1 : nstrings
-
-   for j = 1 : nnodes_or
-
-       if C_s(i,j) == -1 
-
-           sconn_or(i,1) = j;
-
-       end
-
-       if C_s(i,j) == 1 
-
-           sconn_or(i,2) = j;
-
-       end
-
-   end
-
-end
-
-class = 0;                        % class of the tensegrity structure
-nb_nodes = zeros(nnodes_or,1);    % vector of the number of bars connected to each node
-n_virt = 0;                       % total number of virtual nodes
-condflag_id = zeros(nnodes_or,1); % vector of id flag of n0 nodes (nodes not connected to bars)
-nnodes_0=0;                       % number of nodes not connected to bars
-
-for i = 1 : nnodes_or
-
-    for j = 1 : nbars
-    
-        nb_nodes(i,1) = nb_nodes(i,1) + abs(C_b(j,i));
-        
-    end
-    
-    if class < nb_nodes(i,1)
-        
-        class = nb_nodes(i,1);
-        
-    end
-    
-    if nb_nodes(i,1) > 1
-       
-        n_virt = n_virt + nb_nodes(i,1) - 1;
-        
-    end
-    
-    if nb_nodes(i,1) == 0
-       
-        condflag_id(i,1) = 1;
-        nnodes_0 = nnodes_0 + 1;
-        
-    end
-    
-end
-
-nnodes = nnodes_or + n_virt; % total number of nodes (2 * nbars + nnodes_0)
-Nmat_NG = zeros(3,nnodes);      % total matrix of nodal coordinates NAGASE
-
-% vector of nodal flags: 
-% id = 1: phisical node; 
-% id = 2: virtual node; 
-% id = 3: n0 node
-id_node_flags = zeros(nnodes,1);
-id_node_corr = zeros(3,nnodes);  % correspondance between original and new nodes
-
-count = 0;
-count_2 = nnodes;
-for i = 1 : nnodes_or
-    
-    if nb_nodes(i,1) == 0
-                
-        id_node_flags(count_2) = 3;
-        
-        Nmat_NG(:,count_2) = N0(:,i);
-        
-        id_node_corr(1,count_2) = i;
-        id_node_corr(2,count_2) = count_2;
-        
-        count_2 = count_2 - 1; 
-        
-    end
-   
[…]
