### 2019-08-06 e1dd5a5
Delete tenseg_sim_class1open.m
