### 2019-08-06 c1d9c53
Delete tenseg_sim_classkopen.m
