### 2019-08-06 f4aeea6
Delete tenseg_string_segment.m
