### 2019-08-06 bf0224d
Delete tenseg_string_segment.m
