### 2019-08-06 7dc48ea
Delete Main_equilibrium_D_Bar.m
