### 2019-08-06 4232840
Delete tenseg_ind2C.m
