### 2019-08-07 a7da15a
Delete Main_Example_String_Only_Dynamics.m
