### 2019-08-07 2af0f9d
Delete Main_Example_Doubel_Pendulum_Dynamics.m
