### 2019-08-07 843b795
Delete Main_Example_class_k_Dbar_Constraint_Dynamics.m
