### 2019-08-07 536faae
Delete Main_Example_class_k_X.m
