### 2019-08-07 3f0f234
Delete Main_Example_class1_Prism_Dynamics.m
