### 2019-08-07 ae957fa
Delete Example_cross_bar.m
