### 2019-08-07 ef27f5b
Delete Main_Example_equilibrium_kite.m
