### 2019-08-07 9f025a5
Delete Example_equilibrium_prism.m
