### 2019-08-07 cc1f40c
Delete tenseg_plot_physical.m
