### 2019-08-07 3f475fe
Delete Main_Double_Pendulum_Analytical.m
