### 2019-08-07 d28ff18
Delete Main_Doubel_Pendulum_Dynamics_TEAM.m
