+++ Test_Examples/Dynamics/Doubel_Pendulum/README.md
@@ -1,7 +1,0 @@
-This folder gives a comparison of nodes positions and bar length errors of TEAM software and analytical solution by a double pendulum.
-
-The results are shown in **`TEAM_n_Analytical_Solution_Comparison`** folder. 
-
-If the user wants to re-run or check other examples, go and run **`Main_Double_Pendulum_Analytical.m`** and **`Main_Doubel_Pendulum_Dynamics_TEAM.m`** files in **`Analytical_Solution`** and **`TEAM_Solution`**. 
-
-The results will be stored in .mat files. Then, copy and paste the .mat files into **`TEAM_n_Analytical_Solution_Comparison`**, and run **`Main_Error_History.m`**. Please make sure the dimension of t is the same.
