### 2019-08-07 2cbaa54
Delete Main_equilibrium_T_Bar.m
