### 2019-08-07 244a6d3
Delete nonlinear_force_density_minimal_mass.m
