### 2019-08-07 2d7d5d7
Delete tenseg_axisview.m
