### 2019-08-07 01de61c
Delete tenseg_equilibrium_minimal_mass.m
