+++ Test_Examples/Statics/T-bar_Statics/tenseg_equilibrium_minimal_mass.m
@@ -1,222 +1,0 @@
-function [Force_den,Support_Load,LABEL,MIN_MASS,Mass_bar,Mass_string,BarWidth_in,StringWidth_in,Loop] = tenseg_equilibrium_minimal_mass(tenseg_struct)
-% /* This Source Code Form is subject to the terms of the Mozilla Public
-% * License, v. 2.0. If a copy of the MPL was not distributed with this
-% * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
-
-% INPUT: STRUCTURES
-% OUTPUT1: FORCE_DENSITY;
-% OUTPUT2: SUPPORT_LOAD;
-% OUTPUT3: LABEL; IF THE ENTRIES ARE 1, IT'S BUCKLING; OTHERWISE IT'S
-% YIELDING
-
-% Load in structure definitions
-N = tenseg_struct.N;
-C_s = tenseg_struct.C_s;
-C_b = tenseg_struct.C_b;
-W = tenseg_struct.W;
-% bar_failure=tenseg_struct.bar_failure; % 'yielding' or 'yielding_n_buckling'
-bar_material=tenseg_struct.bar_material;
-string_material=tenseg_struct.string_material;
-Pinned_nodes = tenseg_struct.Pinned_nodes;
-
-for i = 1:size(N,2)
-    E = [C_s'*diag(C_s(:,i)) -C_b'*diag(C_b(:,i))];
-    K(3*i-2:3*i,:) = N*E;
-end
-K1 = K;
-W_vec = reshape(W,[],1);
-
-if ~isempty(Pinned_nodes)
-    % Pinned Nodes
-    remove = [];
-    j = 1;
-    nc = Pinned_nodes(:,1)';
-    for i = 1:size(nc,2)
-        if Pinned_nodes(i,2) == 1
-            K2(j,:) = K(3*nc(i)-2,:);
-            remove = [remove 3*nc(i)-2];
-            j = j+1;
-        end
-        if Pinned_nodes(i,3) == 1
-            K2(j,:) = K(3*nc(i)-1,:);
-            remove = [remove 3*nc(i)-1];
-            j = j+1;
-        end
-        if Pinned_nodes(i,4) == 1
-            K2(j,:) = K(3*nc(i),:);
-            remove = [remove 3*nc(i)];
-            j = j+1;
-        end
-    end
-    
-    
-    % Defined Loads
-    K1(remove,:) = [];
-    W_vec(remove,:) = [];
-    
-end
-
-size(K1)
-rank(K1)
-% null(K1)
-% null(K1')
-rank([K1 W_vec])
-
-
-switch bar_material
-    case 'Steel'
-        % -------------- Steel Properties for Bar -----------
-        E_b = 200e09;
-        rho_b = 8000;
-        sigma_b = 300e06;
-    case 'UHMWPE'
-        %---------------- UHMWPE Properties for Bar ----------
-        E_b = 120e09;
-        rho_b = 970;
-        sigma_b = 2.7e09;
-    case 'Aluminum'
-        %---------------- Aluminum for Bar -------------------
-        E_b = 60e09;
-        rho_b = 2700;
-        sigma_b = 110e06;
-    otherwise
-        disp('Edit the material database')
-end
-
-switch string_material
-    case 'Steel'
-        %% -------------- Steel Properties for Strings -----------
-        E_s = 200e09;
-        rho_s = 8000;
-        sigma_s = 300e06;
-    case 'UHMWPE'
-        %------------------ UHMWPE Properties for Strings ---------
[…]
