+++ JOSS_Paper/paper.bib
@@ -1,4 +1,4 @@
-@inproceedings{Goyal_2017_Habitat,
+@inproceedings{GoyalHabitat,
   title={Design and Control of Growth Adaptable Artificial Gravity Space Habitat},
   author={Goyal, Raman and Bryant, Tyler and Majji, Manoranjan and Skelton, Robert E and Longman, Anthony},
   booktitle={AIAA SPACE and Astronautics Forum and Exposition},
@@ -7,7 +7,7 @@
   doi ={10.2514/6.2017-5141}, 
 }
 
-@inproceedings{Chen_2018_Habitat,
+@inproceedings{ChenHabitat,
   title={Energy Analysis of Growth Adaptable Artificial Gravity Space Habitat},
   author={Chen, Muhao and Shen, Yuling and Goyal, Raman  and Majji, Manoranjan and Skelton, Robert E.},
   booktitle={AIAA SPACE and Astronautics Forum and Exposition},
@@ -16,7 +16,7 @@
   doi ={10.2514/6.2018-5109},
 }
 
-@inproceedings{Goyal_2018_ASCE,
+@inproceedings{GoyalASCE,
 title={Dynamics of Class 1 Tensegrity Systems Including Cable Mass},
 author={Goyal, R. and Skelton, R. E.},
 booktitle={16th Biennial ASCE Conference - Earth and Space 2018},
@@ -26,7 +26,7 @@ URL = {https://ascelibrary.org/doi/abs/10.1061/9780784481899.082},
 year={2018}
 }
 
-@article{Karnan_Goyal_2017_IROS, 
+@article{KarnanGoyalIROS, 
 author={H. Karnan and R. Goyal and M. Majji and R. E. Skelton and P. Singla}, 
 journal={2017 IEEE/RSJ-IROS}, 
 title={Visual feedback control of tensegrity robotic systems}, 
@@ -37,14 +37,14 @@ ISSN={2153-0866},
 month={Sept},
 }
 
-@article{Goyal_2019_Buckling,
+@article{GoyalBuckling,
   title={Analytical Study of Tensegrity Lattices for Mass-Efficient Mechanical Energy Absorption},
   author={Goyal, R. and Hernandez,Edwin A. Peraza and Skelton, R. E.},
   journal={International Journal of Space Structures},
   year={2019}
 }
 
-@article{Goyal_Dynamics_2019,
+@article{GoyalDynamics,
   title={Tensegrity System Dynamics with Rigid Bars and Massive Strings},
   author={Goyal, R. and Skelton, R. E.},
   journal={Multibody System Dynamics},
@@ -53,7 +53,7 @@ month={Sept},
   year={2019}
 }
 
-@book{Skelton_2009_Tensegrity_Book,
+@book{SkeltonTensegrityBook,
   title={Tensegrity Systems},
   author={Skelton, R. E. and de Oliveira, M. C.},
   ISBN = {9780387742427},
@@ -61,7 +61,7 @@ month={Sept},
   publisher={Springer US}
 }
 
-@article{Vytas_2013,
+@article{Vytas,
   title={Tensegrity based probes for planetary exploration: Entry, descent and landing ({EDL}) and surface mobility analysis},
   author={SunSpiral, V. and Gorospe, G. and Bruce, J. and Iscen, A. and Korbel, G. and Milam, S. and Agogino, A. and Atkinson, D.},
   journal={International Journal of Planetary Probes},
@@ -69,7 +69,7 @@ month={Sept},
   year={2013}
 }
 
-@inproceedings{Nagase_2014_Minimal,
+@inproceedings{NagaseMinimal,
   title={Minimal mass design of tensegrity structures},
   author={Nagase, K. and Skelton, R. E.},
[…]
