### 2019-08-07 efadb3c
Update paper.md
