+++ JOSS_Paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'TEAM: Tensegrity Engineering Analysis Master'
+title: 'MOTES: Modeling of Tensegrity Structures'
 
 tags:
   - Tensegrity systems
@@ -42,10 +42,9 @@ Most approaches in the field of Multi-body dynamics use a minimal coordinate rep
 
 Moreover, all the strings are also connected to the node centers of the bars, resulting in no torque (no rotation) along the axis of the bar. The use of non-minimal system results in a second-order matrix differential equation to describe the dynamics of any tensegrity system. One disadvantage of such a method is the violation of the rigid body (constant bar length) constraint in the presence of computational errors. In order to satisfy the constant bar length constraint, we use an analytical correction step to eliminate the approximated errors at each integration step. 
 
-Tensegrity Engineering Analysis Master (TEAM) provides two categories for the analysis of any tensegrity structure. First, the static analysis provides the required tension in the strings and force in the bars for a self-equilibrium state or the equilibrium condition in the presence of an external force. This analysis uses \emph{Linear Programming} to solve for the minimum mass required under yielding constraints. The software also allows solving for minimum mass under buckling and yielding failure criteria using \emph{non-linear optimization}. 
+Modeling of Tensegrity Structures (MOTES) provides two categories for the analysis of any tensegrity structure. First, the static analysis provides the required tension in the strings and force in the bars for a self-equilibrium state or the equilibrium condition in the presence of an external force. This analysis uses \emph{Linear Programming} to solve for the minimum mass required under yielding constraints. The software also allows solving for minimum mass under buckling and yielding failure criteria using \emph{non-linear optimization}. 
 
-Second, the dynamic analysis uses a second-order matrix differential equation to simulate the dynamics of any complexity of tensegrity structure. The software runs a self-developed modified Runge-Kutta integration package to solve the nonlinear differential equations. A bar length correction scheme is used to correct the dynamics response that might incur random errors because of computational limitations. This analytical correction step also restricts the errors in connection constraints for class-k structures. The class-k, bar-to-bar connections (ball joints) are formulated as linear constraints. These constraints in the motion space give rise to constraint forces in the structure. An analytical solution is provided to solve for the constraint forces, and a reduced order model is developed to simulate the dynamics for this restricted motion space.
-The mass in the strings is also included in the dynamics by discretizing the string into several point masses along the length of the string. This complete mathematical formulation for the tensegrity dynamics is developed in our recent dynamics paper [@GoyalDynamics].
+Second, the dynamic analysis uses a second-order matrix differential equation to simulate the dynamics of any complexity of tensegrity structure. The software runs a self-developed modified Runge-Kutta integration package to solve the nonlinear differential equations. A bar length correction scheme is used to correct the dynamics response that might incur random errors because of computational limitations. This analytical correction step also restricts the errors in connection constraints for class-k structures. The class-k, bar-to-bar connections (ball joints) are formulated as linear constraints. These constraints in the motion space give rise to constraint forces in the structure. An analytical solution is provided to solve for the constraint forces, and a reduced order model is developed to simulate the dynamics for this restricted motion space. The mass in the strings is also included in the dynamics by discretizing the string into several point masses along the length of the string. This complete mathematical formulation for the tensegrity dynamics is developed in our recent dynamics paper [@GoyalDynamics].
 
 The software is being used to develop various tensegrity structures like Tensegrity robotic arm, Tensegrity antenna, Tensegrity lander, and Space Habitat, where we integrate structure and control design to get the required performance.
 The software was also used as a part of the class curriculum for the course \textit{AERO-489/689 Design Elective: Advanced Statics and Dynamics of Flexible Structures: Tensegrity Systems} which was offered in spring 2019.
