+++ JOSS_Paper/paper.md
@@ -46,7 +46,7 @@ Modeling of Tensegrity Structures (MOTES) provides two categories for the analys
 
 Second, the dynamic analysis uses a second-order matrix differential equation to simulate the dynamics of any complexity of tensegrity structure. The software runs a self-developed modified Runge-Kutta integration package to solve the nonlinear differential equations. A bar length correction scheme is used to correct the dynamics response that might incur random errors because of computational limitations. This analytical correction step also restricts the errors in connection constraints for class-k structures. The class-k, bar-to-bar connections (ball joints) are formulated as linear constraints. These constraints in the motion space give rise to constraint forces in the structure. An analytical solution is provided to solve for the constraint forces, and a reduced order model is developed to simulate the dynamics for this restricted motion space. The mass in the strings is also included in the dynamics by discretizing the string into several point masses along the length of the string. This complete mathematical formulation for the tensegrity dynamics is developed in our recent dynamics paper [@GoyalDynamics].
 
-The software is being used to develop various tensegrity structures like Tensegrity robotic arm, Tensegrity antenna, Tensegrity lander, and Space Habitat, where we integrate structure and control design to get the required performance.
+The software is being used to develop various tensegrity structures like tensegrity robotic arm, tensegrity antenna, tensegrity lander, and space habitat, where we integrate structure and control design to get the required performance.
 The software was also used as a part of the class curriculum for the course \textit{AERO-489/689 Design Elective: Advanced Statics and Dynamics of Flexible Structures: Tensegrity Systems} which was offered in spring 2019.
 
 # Acknowledgements
