### 2019-08-07 a6520d2
Update setup.m
