### 2019-08-07 b4e9cd4
Update Main_Double_Pendulum_Dynamics_MOTES.m
