+++ Software_Verification_and_Examples/Dynamics_Examples/00_Software_Verification_Double_Pendulum/MOTES_Solution/Main_Double_Pendulum_Dynamics_MOTES.m
@@ -91,7 +91,7 @@ classK_test.W= W;
 
 %%
 % Create animation
-% tenseg_animation(History,classK_test,[],[],[],[],10)
+tenseg_animation(History,classK_test,[],[],[],[],10)
 
 %%
 % Plot Node History
