### 2019-08-07 710742f
Update Main_Example_String_Only_Dynamics.m
