+++ Software_Verification_and_Examples/Dynamics_Examples/01_Example_Class_0_String_Only_Dynamics/Main_Example_String_Only_Dynamics.m
@@ -60,7 +60,7 @@ Snet.tf = 5; % Simulation time
 [History,sim_debug] = tenseg_sim_class1open(Snet);
 
 %% Create animation
-% tenseg_animation(History,Snet,[],[],[],[],10)
+tenseg_animation(History,Snet,[],[],[],[],10)
 
 %% Plot History
 
