+++ README.md
@@ -36,7 +36,7 @@ This folder contains the journal paper corresponding to the software, including 
 
 This folder contains:
 
-All the all library functions for tensegrity statics and dynamcis analysis are organized in this folder. By following instructions of statics and dynamics analysis from ''User_Guide.pdf", one can perform the analysis.
+All the library functions for tensegrity statics and dynamcis analysis are organized in this folder. By following instructions of statics and dynamics analysis from ''User_Guide.pdf", one can perform the analysis.
 
 #### Software Verification and Examples
 
@@ -55,7 +55,7 @@ Here, we give examples to verify and demonstrate the statics of this software.
 This folder contains "User_Guide.pdf". The file provides the detailed description, how to use the software, various applications, and how to become a developer as well.
 
 #### Videos
-Some interesting tensegrity animation examples are shown here.
+Some interesting tensegrity animation examples are shown in this folder.
 
 ---
 
