### 2019-08-12 #1613 @ctdegroot
@ramaniitrgoyal92 A few things to look at for now before I get into looking at the actual software:

- It is stated that the version is 2.1, but there are no tags in the repository to indicate this.
- The paper just contains one main section "Summary". I think it would be better if you added a couple of sections to separate the background information from the description of the software.
