### 2019-08-17 b992e53
Update paper.md
