+++ JOSS_Paper/paper.md
@@ -34,20 +34,26 @@ date: 15 June 2019
 bibliography: paper.bib
 ---
 
-# Summary
+# Tensegrity Structures
 
-Tensegrity systems [@SkeltonTensegrityBook] dynamics is a subset of the class of Multi-body dynamics which includes cylindrical rigid bodies (bars) and elastic members (strings) arranged in a stabilizable topology. The name ‘Tensegrity’ is coined by Buckminster Fuller [@Fuller]. The name comes by adding ‘Tensile +Integrity = Tensegrity’ for the art-form created by Ioganson (1921) and Snelson (1948) [@Snelson]. Tensegrity structures are created by methodically arranging tensile members (strings) and compressive members (bars) to form a stable system. Skelton defines "Class-1" tensegrity system where no compressive members touch each other, and a "Class-k" system, where k compressive members are connected at a node [@SkeltonTensegrityBook]. The multiple compressive members are connected through ball joints, causing all the members in a tensegrity structure to be axially loaded, i.e., no moment is present on any individual member. Tensegrity structures can also be prestressed to have uni-directional loading for all the individual members giving the freedom to design tension and compression members separately. This freedom provides good structural efficiency (high strength to mass ratio) to tensegrity structures [@SkeltonTensegrityBook]. A particular topology of tensegrity structure can provide various self-equilibrium solutions corresponding to different values of prestress in the structure. The various stable prestress values provide a domain set to minimize the mass of the structure. This minimization has proved tensegrity to be an optimal mass solution for various loading conditions. Moreover, different prestress values correspond to different stiffness, allowing to change the stiffness without changing the shape. These properties of tensegrity structure have led the researchers to use tensegrity concepts in various applications from Civil engineering bridges [@CarpentieriBridge] to robotic manipulators [@KarnanGoyalIROS] to various space applications; landers [@Vytas], [@GoyalBuckling], deployable structures [@TibertPellegrinoMast], and, space habitat designs [@GoyalHabitat], [@ChenHabitat].
[…]
