+++ JOSS_Paper/paper.bib
@@ -1,4 +1,4 @@
-@inproceedings{GoyalHabitat,
+@inproceedings{Goyal_2017_Habitat,
   title={Design and Control of Growth Adaptable Artificial Gravity Space Habitat},
   author={Goyal, Raman and Bryant, Tyler and Majji, Manoranjan and Skelton, Robert E and Longman, Anthony},
   booktitle={AIAA SPACE and Astronautics Forum and Exposition},
@@ -7,7 +7,7 @@
   doi ={10.2514/6.2017-5141}, 
 }
 
-@inproceedings{ChenHabitat,
+@inproceedings{Chen_2018_Habitat,
   title={Energy Analysis of Growth Adaptable Artificial Gravity Space Habitat},
   author={Chen, Muhao and Shen, Yuling and Goyal, Raman  and Majji, Manoranjan and Skelton, Robert E.},
   booktitle={AIAA SPACE and Astronautics Forum and Exposition},
@@ -16,7 +16,7 @@
   doi ={10.2514/6.2018-5109},
 }
 
-@inproceedings{GoyalASCE,
+@inproceedings{Goyal_2018_ASCE,
 title={Dynamics of Class 1 Tensegrity Systems Including Cable Mass},
 author={Goyal, R. and Skelton, R. E.},
 booktitle={16th Biennial ASCE Conference - Earth and Space 2018},
@@ -26,7 +26,7 @@ URL = {https://ascelibrary.org/doi/abs/10.1061/9780784481899.082},
 year={2018}
 }
 
-@article{KarnanGoyalIROS, 
+@article{Karnan_Goyal_2017_IROS, 
 author={H. Karnan and R. Goyal and M. Majji and R. E. Skelton and P. Singla}, 
 journal={2017 IEEE/RSJ-IROS}, 
 title={Visual feedback control of tensegrity robotic systems}, 
@@ -37,23 +37,39 @@ ISSN={2153-0866},
 month={Sept},
 }
 
-@article{GoyalBuckling,
+@article{Goyal_2019_Buckling,
   title={Analytical Study of Tensegrity Lattices for Mass-Efficient Mechanical Energy Absorption},
   author={Goyal, R. and Hernandez,Edwin A. Peraza and Skelton, R. E.},
   journal={International Journal of Space Structures},
+  pages={1-19},
+  DOI = {10.1177/0956059919845330},
   year={2019}
 }
 
-@article{GoyalDynamics,
+@INPROCEEDINGS{Sabelhaus, 
+author={A. P. {Sabelhaus} and A. K. {Akella} and Z. A. {Ahmad} and V. {SunSpiral}}, 
+booktitle={2017 American Control Conference (ACC)}, 
+title={Model-Predictive Control of a flexible spine robot}, 
+year={2017}, 
+volume={}, 
+number={}, 
+pages={5051-5057}, 
+doi={10.23919/ACC.2017.7963738}, 
+ISSN={2378-5861}, 
+month={May},}
+
+@article{Goyal_Dynamics_2019,
   title={Tensegrity System Dynamics with Rigid Bars and Massive Strings},
   author={Goyal, R. and Skelton, R. E.},
   journal={Multibody System Dynamics},
-  pages = {1-26},
+  volume={46},
+  number={3},
+  pages={203-228},
   DOI = {10.1007/s11044-019-09666-4},
   year={2019}
 }
 
-@book{SkeltonTensegrityBook,
+@book{Skelton_2009_Tensegrity_Book,
   title={Tensegrity Systems},
   author={Skelton, R. E. and de Oliveira, M. C.},
   ISBN = {9780387742427},
@@ -61,7 +77,7 @@ month={Sept},
   publisher={Springer US}
 }
 
-@article{Vytas,
+@article{Vytas_2013,
[…]
