### 2019-08-18 #1613 @vaishnavtv
@ramaniitrgoyal92

The version number as submitted to JOSS for review is 2.1, but the technical manual document lists it as 1.x. It'd be helpful to tag releases from the Github repository itself, to avoid conflicting information.
