### 2019-08-24 #1613 @Kevin-Mattheus-Moerman
> @ctdegroot We did the suggested changes. Thanks a lot!
@ctdegroot can you verify the changes made? Thanks
