### 2019-08-24 #1613 @Kevin-Mattheus-Moerman
> @ramaniitrgoyal92
>
> The version number as submitted to JOSS for review is 2.1, but the technical manual document lists it as 1.x. It'd be helpful to tag releases from the Github repository itself, to avoid conflicting information.

@vaishnavtv you can tick the version box (it will be removed in future review issues) as we fix the version tag on acceptance.
