### 2019-08-24 768fe29
Update Main_Double_Pendulum_Dynamics_MOTES.m
