### 2019-08-24 1daa814
Update tenseg_dyn_c1open_fnc.m
