### 2019-08-24 5c76925
Update tenseg_dyn_ckopen_fnc.m
