### 2019-08-24 382e6d8
Update Main_equilibrium_D_Bar.m
