### 2019-08-24 96ae543
Update tenseg_equilibrium_minimal_mass.m
