### 2019-08-24 89c0f96
Update Main_Example_String_Only_Dynamics.m
