### 2019-08-24 261e029
Update Main_Example_class_k_X.m
