### 2019-08-24 0664544
Update Main_Example_class1_Prism_Dynamics.m
