### 2019-08-24 1860827
Update Example_cross_bar.m
