### 2019-08-24 0082ce8
Update Main_Example_equilibrium_kite.m
