+++ Software_Verification_and_Examples/Statics_Examples/01_Example_equilibrium_kite/Main_Example_equilibrium_kite.m
@@ -8,7 +8,7 @@
 %    - Prestress equilibrium solver
 %    - Pinned_nodes = [n1 1 0 1; n2 1 0 0]; n1 node fixed in x and z direction, n2 node fixed in x direction 
 
-clear all; clc; close all
+clear all; clc; close all; warning off
 
 % Manually specify node positions (in 3D).
 n1 = [-1 1/2 0]';
