### 2019-08-24 9e74736
Update Example_equilibrium_prism.m
