### 2019-08-24 27a14da
Update Main_equilibrium_T_Bar.m
