+++ Software_Verification_and_Examples/Statics_Examples/03_Example_T_bar_Statics/Main_equilibrium_T_Bar.m
@@ -9,7 +9,7 @@
 %    - Prestress equilibrium solver
 %    - Pinned_nodes = [n1 1 0 1; n2 1 0 0];
 %      n1 node fixed in x and z direction, n2 node fixed in x direction
-clear all; clc; close all;
+clear all; clc; close all; warning off;
 % Manually specify node positions (in 3D).
 n1 = [-1 0 0]';
 n2 = [0 -1 0]';
