### 2019-09-03 #1613 @apsabelhaus
@Kevin-Mattheus-Moerman, it looks like @ramaniitrgoyal92 has changed the name of the repository / submission (from TEAM to MOTES.) I've checked the "code available at the url" box, since Github redirects to the new repository, but it seems the original URL is to the "TEAM" name. Is there a way that the authors should be updating the JOSS submission with the new name, etc.? Are there other dependencies based on name that need to be looked at by the reviewers?
