### 2019-09-03 #1613 @apsabelhaus
@ramaniitrgoyal92 I went through a second round of review just now and opened some more issues. Generally speaking, it seems like this review is taking a while, and I don't think that's just because it's the summertime. The code and documentation are still difficult to read and understand, contain typos, etc. This review would probably go by faster if y'all took a second look at JOSS' standards and verified how you think you meet them all, maybe by responding in the issues with an explanation. Additionally, like I recommended in a response to an issue, describing your changes specifically is better than "thanks, we fixed it!" since that doesn't tell me what you changed.
