+++ JOSS_Paper/paper.md
@@ -41,9 +41,9 @@ Tensegrity systems [@Skelton_2009_Tensegrity_Book] dynamics is a subset of the c
 
 # Software Description
 
-Modeling of Tensegrity Structures (MOTES) provides two categories for the analysis of any tensegrity structure. First, the static analysis provides the minimum mass of the tensegrity structure by optimizing for the tensile force in the strings and compressive force in the bars for no external force (self-equilibrium state) and in the presence of a given external force. The optimization problem is written as a \emph{Linear Programming} to solve for the minimum mass required under yielding constraints. The software also allows to solve for the minimum mass under buckling and yielding failure criteria through a \emph{non-linear optimization} solver. Second, the dynamic analysis uses a second-order matrix differential equation to simulate the dynamics of any complexity of tensegrity structure [@Goyal_Dynamics_2019]. This dynamic model assumes the bars to be rigid and strings to show elastic behaviour (Hookean).
+Modeling of Tensegrity Structures (MOTES) provides two categories for the analysis of any tensegrity structure. First, the static analysis provides the minimum mass of the tensegrity structure by optimizing for the tensile force in the strings and compressive force in the bars for no external force (self-equilibrium state) and in the presence of a given external force. The optimization problem is written as a \emph{Linear Programming} to solve for the minimum mass required under yielding constraints. The software also allows solving for the minimum mass under buckling and yielding failure criteria through a \emph{non-linear optimization} solver. Second, the dynamic analysis uses a second-order matrix differential equation to simulate the dynamics of any complexity of tensegrity structure [@Goyal_Dynamics_2019]. This dynamic model assumes the bars to be rigid and strings to show elastic behavior (Hookean).
 
-The software runs a self-developed modified Runge-Kutta integration package to solve the nonlinear differential equations. A bar length correction scheme is used to correct the dynamics response that might incur random errors because of computational limitations. This analytical correction step also restricts the errors in connection constraints for class-k structures. The class-k, bar-to-bar connections (ball joints) are formulated as linear constraints. These constraints in the motion space give rise to constraint forces in the structure. An analytical solution is provided to solve for the constraint forces, and a reduced order model is developed to simulate the dynamics in the restricted motion space.
+The software runs a self-developed modified Runge-Kutta integration package to solve the nonlinear differential equations. A bar length correction scheme is used to correct the dynamics response that might incur random errors because of computational limitations. This analytical correction step also restricts the errors in connection constraints for class-k structures. The class-k, bar-to-bar connections (ball joints) are formulated as linear constraints. These constraints in the motion space give rise to constraint forces in the structure. An analytical solution is provided to solve for the constraint forces, and a reduced-order model is developed to simulate the dynamics in the restricted motion space.
 The mass in the strings is also included in the dynamics by discretizing the string into several point masses along the length of the string. This complete mathematical model for the tensegrity dynamics is developed in our recent dynamics paper [@Goyal_Dynamics_2019].
 
 # Advantages and Applications
