+++ Function_Library/ode4Ck.m
@@ -1,4 +1,7 @@
 function Y = ode4Ck(odefun,tspan,y0,varargin)
+
+% This function is modified from the ode4.m by Kelly Kearney: https://github.com/kakearney/ecosystem-pkg/blob/master/odefixed/ode4.m. 
+
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
