+++ README.md
@@ -6,6 +6,10 @@
  
 Our research team focuses on integrating structure and control design using Tensegrity structure. The group focuses on designing tensegrity structures to meet the specified objectives. These objectives can vary from minimizing the mass of the structure to controlling the structure to meet specific performance. This software is intended to study the statics and dynamics of tensegrity systems. The authors would like to make this as open-source software to help other researchers who are also interested in this field. 
 
+Contribution of this software is mainly in these two aspects: First, the static analysis provides the minimum mass of the tensegrity structure by optimizing for the tensile force in the strings and compressive force in the bars for no external force (self-equilibrium state) and in the presence of a given external force. The optimization problem is written as a \emph{Linear Programming} to solve for the minimum mass required under yielding constraints. The software also allows solving for the minimum mass under buckling and yielding failure criteria through a \emph{non-linear optimization} solver. 
+Second, the dynamic analysis uses a second-order matrix differential equation to simulate the dynamics of any complexity of tensegrity structure. This dynamic model assumes the bars to be rigid and strings to show elastic behavior (Hookean).
+
+
 #### LICENSE
 
     /* This Source Code Form is subject to the terms of the Mozilla Public
