### 2019-09-08 ea29330
Update PlotCylinderObject.m
