### 2019-09-08 f0a4c6c
Update PlotCylinderObject.m
