### 2019-09-08 6e44cad
Update bar_length_correction_Class1.m
