### 2019-09-08 722c7ec
Update bar_length_correction_ClassK.m
