### 2019-09-08 f298a49
Update extractSignals.m
