+++ Function_Library/extractSignals.m
@@ -2,7 +2,7 @@ function signals = extractSignals(odefun, names)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+%
 %EXTRACTSIGNALS Get log signals from ODE output.
 %   SIGNALS = EXTRACTSIGNALS(@ODEFUN, NAMES) evaluates the ODEFUN to get the
 %       auxiliary output signal logs, and assigns them to a structure with
@@ -12,9 +12,10 @@ function signals = extractSignals(odefun, names)
 %   as the signals in the log output. If more NAMES are provided than
 %   signals in the log cell produced by ODEFUN, the additional names will
 %   be ignored.
+%
 
-	% Get ODE output. Assumes that there are three ODEFUN inputs: time,
-    % state, and parameters.
+% Get ODE output. Assumes that there are three ODEFUN inputs: time,
+% state, and parameters.
     [~, logsout] = odefun([], [], []);
 
     % Get number of signals and number of named signals.
