### 2019-09-08 90e8aee
Update tenseg_animation.m
