### 2019-09-08 de93a75
Update tenseg_class_k_convert.m
