### 2019-09-08 fa29b9a
Update tenseg_defaults.m
