+++ Function_Library/tenseg_defaults.m
@@ -2,7 +2,7 @@ function default = tenseg_defaults(n,beta,alpha,sigma)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+%
 % default = TENSEG_DEFAULTS(n,beta,alpha,sigma) creates a structure
 % containing all default variable specifications in the event that the
 % variable is not explicitly specified by the user
