### 2019-09-08 ee4bdfc
Update tenseg_dyn_c1open_fnc.m
