### 2019-09-08 810ab4a
Update tenseg_dyn_ckopen_fnc.m
