### 2019-09-08 6a8fd0b
Update tenseg_equilibrium.m
