### 2019-09-08 4d9b09f
Update tenseg_field_init.m
