### 2019-09-08 e9a1afb
Update tenseg_ind2C.m
