### 2019-09-08 d581258
Update tenseg_merge.m
