### 2019-09-08 cd0ebdb
Update tenseg_plot_configurations.m
