### 2019-09-08 2884ba0
Update tenseg_plot_node.m
