### 2019-09-08 f5b0d06
Update tenseg_plot_velocity.m
