### 2019-09-08 b12cb3b
Update tenseg_reordering_topology.m
