+++ Function_Library/tenseg_reordering_topology.m
@@ -1,7 +1,7 @@
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+%
 % Re-orders the node indeces if they are not ordered correctely.
 % This makes the bars connectivity matrix equal to [-I I 0] in the general
 % case
