### 2019-09-08 fea123c
Update tenseg_sim_class1open.m
