### 2019-09-08 e2e9284
Update tenseg_sim_classkopen.m
