### 2019-09-08 03dfd40
Update tenseg_string_segment.m
