### 2019-09-08 e2d899b
Update tenseg_struct_gen.m
