### 2019-09-08 f8979f8
Update Main_Example_String_Only_Dynamics.m
