+++ Software_Verification_and_Examples/Dynamics_Examples/01_Example_Class_0_String_Only_Dynamics/Main_Example_String_Only_Dynamics.m
@@ -1,9 +1,9 @@
 %% Example: String_Only_Dynamics
-
+%
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+%
 % EXAMPLE:
 %    - Manual node matrix specification
 %    - Manual connectivity matrix generation
