### 2019-09-08 1a6adcb
Update Main_Example_equilibrium_kite.m
