+++ Software_Verification_and_Examples/Statics_Examples/01_Example_equilibrium_kite/Main_Example_equilibrium_kite.m
@@ -1,7 +1,7 @@
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+%
 % EXAMPLE:
 %    - Manual node matrix specification
 %    - Manual connectivity matrix generation
