### 2019-09-08 bb7fc8b
Update Example_equilibrium_prism.m
