### 2019-09-08 61aeb59
Update tenseg_string_segment.m
