### 2019-09-08 614afd2
Update tenseg_equilibrium_minimal_mass.m
