+++ Software_Verification_and_Examples/Statics_Examples/00_Software_Verification_D_bar_Statics/tenseg_equilibrium_minimal_mass.m
@@ -2,8 +2,7 @@ function [Force_den,Support_Load,LABEL,MIN_MASS,Mass_bar,Mass_string,BarWidth_in
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
-
+%
 % INPUT: STRUCTURES
 % OUTPUT1: FORCE_DENSITY;
 % OUTPUT2: SUPPORT_LOAD;
