### 2019-09-08 8dba264
Update Main_equilibrium_D_Bar.m
