+++ Software_Verification_and_Examples/Statics_Examples/00_Software_Verification_D_bar_Statics/Main_equilibrium_D_Bar.m
@@ -2,13 +2,14 @@
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+%
 % EXAMPLE:
 %    - Manual node matrix specification
 %    - Manual connectivity matrix generation
 %    - Prestress equilibrium solver
 %    - Pinned_nodes = [n1 1 0 1; n2 1 0 0];
 %      n1 node fixed in x and z direction, n2 node fixed in x direction
+
 clear all; clc; close all; warning off;
 % Manually specify node positions (in 3D).
 n1 = [-1 0 0]';
