### 2019-09-08 e3d3a34
Update tenseg_plot_physical.m
