### 2019-09-08 cf8fefa
Update nonlinear_force_density_minimal_mass.m
