### 2019-09-08 3afa772
Update tenseg_ind2C.m
