+++ Software_Verification_and_Examples/Statics_Examples/03_Example_T_bar_Statics/tenseg_ind2C.m
@@ -2,7 +2,7 @@ function [ C_mat ] = tenseg_ind2C( C_ind,Nodes )
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
-
+%
 % [ C_mat ] = TENSEG_IND2C( C_ind, Nodes ) creates a connectivity matrix
 % from input index notation array and node matrix.
 %
