### 2019-09-08 150bf84
Update tenseg_plot_physical.m
