### 2019-09-08 3dd6c84
Delete tenseg_equilibrium_minimal_mass.m
