### 2019-09-08 6832e51
Delete tenseg_ind2C.m
