### 2019-09-08 8a04d9b
Delete tenseg_string_segment.m
