### 2019-09-08 185352a
Delete tenseg_plot_physical.m
