### 2019-09-08 5ece607
Delete nonlinear_force_density_minimal_mass.m
