### 2019-09-08 4fe6932
Delete tenseg_plot_physical.m
