### 2019-09-09 cce5086
Update setup.m
