+++ setup.m
@@ -1,6 +1,6 @@
 %% ^_^ Welcome to Modeling of Tensegrity Structrues(MOTES) software! ^_^ %%
 % SETUP file to be run only the first time
-
+%
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
