+++ Function_Library/skew.m
@@ -1,5 +1,9 @@
 function S = skew(w)
 %
+% /* This Source Code Form is subject to the terms of the Mozilla Public
+% * License, v. 2.0. If a copy of the MPL was not distributed with this
+% * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
+%
 % The function generates a skew-symmetric matrix for a give vector v
 %	S = skew(v)
 
