### 2019-09-29 #1613 @ramaniitrgoyal92
> @ramaniitrgoyal92, sorry for the delay in response, but I think the reviewers would appreciate it if you could comment/respond to the questions in the corresponding issues in the repository itself.

We have responded to the comments in the corresponding issues of the repository.
Thanks!
