### 2019-09-29 #1613 @ramaniitrgoyal92
> [Respond_to_JOSS.docx](https://github.com/openjournals/joss-reviews/files/3606347/Respond_to_JOSS.docx)

The complete list of updates can be found in the attached word document.
