+++ Function_Library/tenseg_plot_physical.m
@@ -1,0 +1,279 @@
+function [fig_out] = tenseg_plot_physical( N,C_b,C_s,fig_handle,highlight_nodes,view_vec, PlotTitle, R3Ddata,Yld_n_Bkl,BarWidth_in,StringWidth_in)
+% /* This Source Code Form is subject to the terms of the Mozilla Public
+% * License, v. 2.0. If a copy of the MPL was not distributed with this
+% * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
+%
+% This function added thickness display apart from tenseg_plot.m
+% Thickness of bars and strings are proportional to the force densities in
+% them.
+%
+% [fig_out] = TENSEG_PLOT( N,C_b,C_s,fig_handle,highlight_nodes,view_vec )
+% creates a rough visualization figure for a given tensegrity structure
+%
+% Inputs:
+%	N: node matrix (3 x n array for n nodes)
+%	C_b (optional): bar connectivity matrix (beta x n array for beta bars)
+%	C_s: string connectivity matrix (alpha x n array for alpha strings)
+%
+%   LABEL: Gives the information on bar failure mode; 11/23/18
+%
+%	fig_handle (optional): figure in which to plot
+%	highlight_nodes (optional): node(s) to highlight when plotting (vector
+%		containing node numbers)
+%	view_vec (optional): plotting view direction (see view())
+%   PlotTitle (optional): Title of the plot, default: ''
+%   R3Ddata (optional): Structure with the radius of objects for 3D plots
+%        [].Bradius: Radius of bars [# bars x 1]
+%        [].Sradius: Radius of strings [# strings x 1]
+%        [].Nradius: Radius of node spheres [# nodes x 1]
+%   Yld_n_Bkl: Yielding or Buckling label
+%   BarWidth_in: line width of bars
+%   StringWidth_in: line width of strings
+% Outputs:
+%	fig_out: figure handle in which plot was made
+%
+% Example:
+%	tenseg_plot_physical( N,C_b,C_s,fig_handle,highlight_nodes,view_vec, PlotTitle, R3Ddata,Yld_n_Bkl,BarWidth_in,StringWidth_in)
+
+% Handle optional arguments
+
+%% Object size options (for line plots)
+% BarWidth = BarWidth_in; % Width of bar lines
+% StringWidth = StringWidth_in ; % Width of string lines
+NodeSize = 20; % Size of node marker
+
+%% Labeling options
+% Write labels? (1: show, 0: suppress)
+LabelNodes = 1;
+LabelStrings = 1;
+LabelBars = 1;
+
+FontBars = 15; % Font of bar labels
+FontStrings = 10; % Font of string labels
+FontNodes = 18; % Font of node labels
+
+FractionDistance = 0.005; % Distance between object and label (relative to overall axis size)
+
+%% 3D plot options
+nsurfpatches = 6; % Number of surface patches in 3D plots
+BarSurfColor = [0.2, 0.2, 0.6];
+StringSurfColor = [0.9, 0.1, 0.1];
+LightAmbientStrength = 0.7;% [0,1], 0.3 is Matlab's default
+
+%%
+switch nargin
+    case 5
+        view_vec = [];
+    case 4
+        view_vec = [];
+        highlight_nodes = [];
+    case 3
+        view_vec = [];
+        highlight_nodes = [];
+        fig_handle = [];
+end
+
+if nargin < 7 % Empty Plot Title
[…]
