### 2019-09-29 6826c01
Update Main_Example_equilibrium_kite.m
