+++ Function_Library/skew.m
@@ -1,4 +1,4 @@
-function S = skew(w)
+function S = skew(v)
 %
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
