+++ Software_Verification_and_Examples/Dynamics_Examples/00_Software_Verification_Double_Pendulum/README.md
@@ -2,6 +2,6 @@ This folder gives a comparison of nodes positions and bar length errors of MOTES
 
 The results are shown in **`MOTES_n_Analytical_Solution_Comparison`** folder. 
 
-If the user wants to re-run or check other examples, go and run **`Main_Double_Pendulum_Analytical.m`** and **`Main_Double_Pendulum_Dynamics_TEAM.m`** files in **`Analytical_Solution`** and **`MOTES_Solution`**. 
+If the user wants to re-run or check other examples, go and run **`Double_Pendulum_Analytical.m`** and **`Double_Pendulum_Dynamics_TEAM.m`** files in **`Analytical_Solution`** and **`MOTES_Solution`**. 
 
-The results will be stored in .mat files. Then, copy and paste the .mat files into **`MOTES_n_Analytical_Solution_Comparison`**, and run **`Main_Error_History.m`**. Please make sure the dimension of time t is the same.
+The results will be stored in .mat files. Then, copy and paste the .mat files into **`MOTES_n_Analytical_Solution_Comparison`**, and run **`Error_History.m`**. Please make sure the dimension of time t is the same.
