### 2019-09-29 ba721e5
Delete Main_Example_cross_bar.m
