+++ JOSS_Paper/paper.bib
@@ -132,3 +132,13 @@ month={May},}
   pages = {1978},
   doi={10.2514/6.2003-1978}
 }
+
+@article{Stedy_2019,
+  title={ STEDY: Software for TEnsegrity DYnamics},
+  author={Vaishnav Tadiparthi, Shao-Chen Hsu, and Raktim Bhattacharya},
+  journal={Journal of Open Source Software},
+  year={2019},
+  volume={4(33)},
+  pages = {1042},
+  doi={10.21105/joss.01042}
+}
