+++ JOSS_Paper/paper.bib
@@ -134,7 +134,7 @@ month={May},}
 }
 
 @article{Stedy_2019,
-  title={ STEDY: Software for TEnsegrity DYnamics},
+  title={STEDY: Software for TEnsegrity DYnamics},
   author={Vaishnav Tadiparthi, Shao-Chen Hsu, and Raktim Bhattacharya},
   journal={Journal of Open Source Software},
   year={2019},
