+++ JOSS_Paper/paper.bib
@@ -139,6 +139,6 @@ month={May},}
   journal={Journal of Open Source Software},
   year={2019},
   volume={4(33)},
-  pages = {1042},
+  pages={1042},
   doi={10.21105/joss.01042}
 }
