+++ JOSS_Paper/paper.md
@@ -49,7 +49,7 @@ The software runs a modified Runge-Kutta integration package to solve the nonlin
 
 Most approaches in the field of Multi-body dynamics use a minimal coordinate representation. This eliminates redundant variables as the body coordinates are used based on connection, from the first body to the last body in a particular order. One disadvantage of this approach is the use of transcendental functions, which can cause significant computational errors. The other disadvantage is the limited configuration of the rigid body connections to a topological tree. The proposed formulation uses a non-minimal coordinate system to describe the dynamics of the system. The approach uses 6 DOF to define a rigid symmetric cylindrical bar with no rotation about the longitudinal axis. For a "Class-k" tensegrity structure, there would be no moment or torque along the bar axis as the bars are connected by frictionless ball joints. 
 
-Moreover, all the strings are also connected to the node centers of the bars, resulting in no torque (no rotation) along the axis of the bar. The use of non-minimal system results in a second-order matrix differential equation to describe the dynamics of any tensegrity system. One disadvantage of such a method is the violation of the rigid body (constant bar length) constraint in the presence of computational errors. In order to satisfy the constant bar length constraint, we use an analytical correction step to eliminate the approximated errors at each integration step. STEDY [@stedy_2019], a similar tensegrity simulator, uses a different correction step and formulation to run the dynamics. The added advantage with MOTES is the ability to model massive strings and to design minimum mass structures using static analysis.
+Moreover, all the strings are also connected to the node centers of the bars, resulting in no torque (no rotation) along the axis of the bar. The use of non-minimal system results in a second-order matrix differential equation to describe the dynamics of any tensegrity system. One disadvantage of such a method is the violation of the rigid body (constant bar length) constraint in the presence of computational errors. In order to satisfy the constant bar length constraint, we use an analytical correction step to eliminate the approximated errors at each integration step. STEDY [@Stedy_2019], a similar tensegrity simulator, uses a different correction step and formulation to run the dynamics. The added advantage with MOTES is the ability to model massive strings and to design minimum mass structures using static analysis.
 
 The software is being used to develop various tensegrity structures like Tensegrity robotic arm, Tensegrity antenna, Tensegrity lander, and Space Habitat, where we integrate structure and control design to get the required performance.
 The software was also used as a part of the class curriculum for the course '\textit{AERO-489/689 Design Elective: Advanced Statics and Dynamics of Flexible Structures: Tensegrity Systems}' which was offered in spring 2019.
