### 2019-10-08 #1613 @apsabelhaus
@ramaniitrgoyal92, you mentioned that you'd responded to comments in the issues in the repository in response to @vaishnavtv's point, but I still had to sort through the word document for issue #1 since the full responses weren't actually in the Github issues. This review is becoming tiresome, and it would be very much appreciated if the authors were more careful and more thorough in the future when saying what you did then actually doing so.
