### 2019-10-08 #1613 @ramaniitrgoyal92
@apsabelhaus Thanks for your comments! I added a note to the #1 issue. (Note: setup.m must be run every time MATLAB opens, before any other file.) Thanks for reminding us and sorry indeed for the inconvenience!
