### 2019-10-08 #1613 @apsabelhaus
@Kevin-Mattheus-Moerman I finished up this penultimate round of reviews. The only remaining issue (which may be substantial, I'm unsure) is the licensing where the authors use someone else's code. If you could take over and work with the authors on that, it would be appreciated, since I can't speak to JOSS' policies. Once that's fixed, I would recommend this submission for publication!
