+++ Function_Library/ode4.m
@@ -1,5 +1,26 @@
 function Y = ode4(odefun,tspan,y0,varargin)
 % This function is written by Kelly Kearney: https://github.com/kakearney/ecosystem-pkg/blob/master/odefixed/ode4.m. 
+% Here is their License 
+%The MIT License (MIT)
+%
+%Copyright (c) 2015 Kelly Kearney
+%
+%Permission is hereby granted, free of charge, to any person obtaining a copy of
+%this software and associated documentation files (the "Software"), to deal in
+%the Software without restriction, including without limitation the rights to
+%use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+%the Software, and to permit persons to whom the Software is furnished to do so,
+%subject to the following conditions:
+%
+%The above copyright notice and this permission notice shall be included in all
+%copies or substantial portions of the Software.
+%
+%THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+%IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+%FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+%COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+%IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+%CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 %
 %ODE4  Solve differential equations with a non-adaptive method of order 4.
 %   Y = ODE4(ODEFUN,TSPAN,Y0) with TSPAN = [T1, T2, T3, ... TN] integrates 
