### 2019-10-08 #1613 @ctdegroot
@Kevin-Mattheus-Moerman I'm satisfied with the changes made by the authors and think that is suitable for all points on the checklist.
