### 2019-10-11 #1613 @Kevin-Mattheus-Moerman
Thanks @apsabelhaus. I can take over. However you also have a "Functionality" box unticked here, are there any remaining issues with the functionality?

In relation to the remaining license issue. Can you explain that in more detail? Are dependencies used without referring to the original works and without including the licences? Or are license conditions from the dependencies not met?
