### 2019-10-11 #1613 @apsabelhaus
@Kevin-Mattheus-Moerman thanks for catching the box-checking mistake - just switched "functionality" with "license."

Summary of the potential licensing problem: (questions for you in parentheses)

- Authors included multiple files from other works and other authors. (Does JOSS approve of including other open-source code in a submission, as long as the original licenses are included, and the code from other sources does not represent a substantial contribution of the submission?)

- Those files originally contained neither attribution nor license from the original source, so were presumably submitted as the authors' original work. Aside - does JOSS have an academic dishonesty statement somewhere that could be referenced in the future?

- This has been corrected, now the files I identified have both attribution and the original license copied into the comments in the file. The authors have reviewed the rest of the repository and they believe that all other code is their original work.

- However, these licenses are different: Apache in the LICENSE file for the authors' code, MIT for the outside code. (Does there need to be some more explanation of which licenses apply where, or is it generally assumed that any license present in an individual file supersedes one in a root directory of a repository?)

That's all! If these answers are "Yes, this is all OK", then I'll have completed the review!
