### 2019-10-19 #1613 @Kevin-Mattheus-Moerman
@ramaniitrgoyal92 can you verify that you clearly state all the third party works and dependencies included with your software? Also can you make sure you check the license terms for these? E.g. if you have to include the license you do etc.

@apsabelhaus thanks for the thorough review! And also for raising these points. If works from others are included they need to be labelled as such (attribution) and the license terms and conditions of such works need to be respected/adhered to.
The license chosen by the author would apply to their work only, not the third party works in the repositoty. If included dependencies have a different license these need to be included. Good practice is to have separate folders for each differently licensed object. Quite often there are no issues with such licenses. The copy-left licenses are an exception as they might force the repo to also carry that license. I believe there are no issues with regards to having MIT licensed content (provided attribution is given and also the license is included) in an Apache-2 repo.
