### 2019-10-21 #1613 @ramaniitrgoyal92
> @ramaniitrgoyal92 can you verify that you clearly state all the third party works and dependencies included with your software? Also can you make sure you check the license terms for these? E.g. if you have to include the license you do etc.
>

@Kevin-Mattheus-Moerman
We have included all the third-party dependencies used in our software. We have added license and attribution to the used functions. We have also added the following in the readme file present in the Function_Library.

"This file contains all necessary functions of tensegrity dynamics and statics.
The ode solver is written by Kelly Kearney: https://github.com/kakearney/ecosystem-pkg/blob/master/odefixed/ode4.m. Their code MIT License (MIT) is included in the functions."

Thank you all for the helpful suggestions!
