+++ JOSS_Paper/paper.md
@@ -36,7 +36,7 @@ bibliography: paper.bib
 
 # Tensegrity Structures
 
-Tensegrity systems [@Skelton_2009_Tensegrity_Book] dynamics is a subset of the class of Multi-body dynamics which includes cylindrical rigid bodies (bars) and elastic members (strings) arranged in a stabilizable topology. The name ‘Tensegrity’ is coined by Buckminster Fuller [@Fuller_1959]. The name comes by adding ‘Tensile +Integrity = Tensegrity’ for the art-form created by Ioganson (1921) and Snelson (1948) [@Snelson_1965].  Tensegrity structures are created by methodically arranging tensile members (strings) and compressive members (bars) to form a stable system. Skelton defines "Class-1" tensegrity system where no compressive members touch each other, and a "Class-$k$" system, where $k$ number of compressive members are connected at a node [@Skelton_2009_Tensegrity_Book]. The multiple compressive members are connected through ball joints, causing all the members in a tensegrity structure to be axially loaded, i.e., no moment is present on any individual member. Tensegrity structures can also be prestressed to have uni-directional loading for all the individual members giving the freedom to design tension and compression members separately. This freedom provides good structural efficiency (high strength-to-mass ratio) to tensegrity structures [@Skelton_2009_Tensegrity_Book]. A particular topology of tensegrity structure can provide various self-equilibrium solutions corresponding to different values of prestress in the structure. The various stable prestress values provide a domain set to minimize the mass of the structure. This minimization has proved tensegrity to be an optimal mass solution for various loading conditions. Moreover, different prestress values correspond to different stiffness, allowing to change the stiffness without changing the shape. These properties of tensegrity structure have led the researchers to use tensegrity concepts in various applications from Civil engineering bridges [@Carpentieri_2015_Bridge] to soft-robotics [@Sabelhaus, @Karnan_Goyal_2017_IROS] to various space applications; landers [@Vytas_2013, @Goyal_2019_Buckling], deployable structures [@Tibert_Pellegrino_2003_Mast], and, space habitat designs [@Goyal_2017_Habitat].
[…]
