+++ JOSS_Paper/paper.bib
@@ -103,7 +103,7 @@ month={May},}
   number={3-4},
   pages={221--243},
   year={2015},
-  doi={10.1260/0266-3511.30.3-4.221}
+  doi={10.1260/0266-3511.30.3-4.221},
   publisher={SAGE Publications Sage UK: London, England}
 }
 
