### 2019-10-21 #1613 @Kevin-Mattheus-Moerman
@ramaniitrgoyal92 Some points on your paper below (I read the version just before you regenerated the pdf, can you tell me what changes you made?). Thanks

- [ ] In `Tensegrity systems (Skelton & Oliveira, 2009) dynamics...` This reads better if the reference is moved after the word `dynamics`. Please make this change if you agree.
- [ ] In `....of the class of Multi-body dynamics...`, does multi-body need the capital letter? consider using a lower case letter. If you chose to use a lower case check your paper throughout.
- [ ] Please use past tense for `The name ‘Tensegrity’ is coined by...`, should be  `The name ‘Tensegrity’ was coined by...`
- [ ] Please rephrase: "The name ‘Tensegrity’ is coined by Buckminster Fuller (Fuller, 1959). The name comes by adding ‘Tensile +Integrity = Tensegrity’ for the art-form created by Ioganson (1921) and Snelson (1948) (Snelson, 1965).",
I propose (unless I misunderstood): "The term tensegrity was first coined by Buckminster Fuller (Fuller, 1959), and is a portmanteau of "tensional integrity. Tensegrity structures also appeared in earlier art works by Ioganson (1921) and Snelson (1948) (Snelson, 1965)."
- [ ] Please rephrase "Skelton defines “Class-1” tensegrity system where no compressive members touch each other, and a “Class-k” system, where k number of compressive members are connected at a node Skelton & Oliveira, 2009).", I propose (unless I misunderstood): "Skelton defines a tensegrity structure as a“Class-1” system if none of the compressive members are connected, if on the other hand k compressive members are connected at a node, this is referred to as a “Class-k” system (Skelton & Oliveira, 2009)."
- [ ] Please change `... allowing to change the stiffness without changing the shape.` to `... allowing one to change the stiffness without changing the shape.`
- [ ] Please change `These properties of tensegrity structure have led the researchers to use tensegrity concepts in various applications from...` to :  `These properties of tensegrity structures have led researchers to use tensegrity concepts in various applications from...`
- [ ] Please rephrase: `First, the static analysis provides the minimum mass of the tensegrity structure by optimizing for the tensile force in the strings and compressive force in the bars for no external force (self-equilibrium state) and in the presence of a given external force. The optimization problem is written as a Linear Programming to solve for the minimum mass required under yielding constraints.`
e.g. to `Firstly, static analysis provides the minimum mass of the tensegrity structure by optimizing tensile forces in the strings and compressive forces in the bars in the absence of external forces (self-equilibrium state), and in the presence of a given external forces.`
- [ ] Please rephrase: `The optimization problem is written as a Linear Programming to solve for the minimum mass required under yielding constraints.` in particular `as a linear programming` sounds off. Should this be something like the following: `In order to solve for the minimum mass required under yielding constraints, MOTES formulates the optimization problem as a "Linear Programming" problem.`
- [ ] Use `Secondly` instead of `Second` in "Second, the dynamic analysis"
- [ ] Can you rephrase this `... and strings to show elastic behavior (Hookean)` (if you agree) to `...and strings to exhibit linear elastic behaviour`.
- [ ] Should `...the dynamics response...` be `...the dynamic response...`?
- [ ] Instead of `The other disadvantage... `, i would use `Another disadvantage...`
- [ ] Explain the acronym DOF
- [ ] Change this `The use of non-minimal system results in...` to something like `The use of the non-minimal system results in...` or `The use of non-minimal systems results in...`
- [ ] `...the approximated errors at each...` should probably be `...the approximation errors at each...`?
- [ ] Please rephrase `...Tensegrity robotic arm, Tensegrity antenna, Tensegrity lander, and Space Habitat, where we integrate structure and control design to get the required performance.` to `...tensegrity robotic
arms, tensegrity antennae, tensegrity landers, and space habitats, where we integrate structure and control design to get the required performance.`
- [ ] When referring to class material at the end of the paper, can you add the university and discipline?
