### 2019-10-21 #1613 @ramaniitrgoyal92
> @ramaniitrgoyal92 Some points on your paper below (I read the version just before you regenerated the pdf, can you tell me what changes you made?). Thanks

@Kevin-Mattheus-Moerman
I added the DOI for the paper title: 'Minimum mass and optimal complexity of planar tensegrity bridges' and added some missing spaces in between references of the first paragraphs of the paper. They were not coming out fine.
