### 2019-10-21 #1613 @ramaniitrgoyal92
>In Tensegrity systems (Skelton & Oliveira, 2009) dynamics... This reads better if the reference is moved after the word dynamics. Please make this change if you agree.

Agreed and changed. Working on the rest.
