### 2019-10-21 c02fa06
Update paper.md
