### 2019-10-21 #1613 @ramaniitrgoyal92
> * [ ]  In `....of the class of Multi-body dynamics...`, does multi-body need the capital letter? consider using a lower case letter. If you chose to use a lower case check your paper throughout.

Multi-body has been replaced with 'multi-body' throughout the paper.

> * [ ]  Please use past tense for `The name ‘Tensegrity’ is coined by...`, should be  `The name ‘Tensegrity’ was coined by...

Now the sentence read:
"The term tensegrity was coined by..."

> * [ ]  Please rephrase: "The name ‘Tensegrity’ is coined by Buckminster Fuller (Fuller, 1959). The name comes by adding ‘Tensile +Integrity = Tensegrity’ for the art-form created by Ioganson (1921) and Snelson (1948) (Snelson, 1965).",
>   I propose (unless I misunderstood): "The term tensegrity was first coined by Buckminster Fuller (Fuller, 1959), and is a portmanteau of "tensional integrity. Tensegrity structures also appeared in earlier art works by Ioganson (1921) and Snelson (1948) (Snelson, 1965)."

The paper now reads:
"The term tensegrity was first coined by Buckminster Fuller (Fuller, 1959), and is a portmanteau of "tensional integrity". Tensegrity structures also appeared in earlier artworks by Ioganson (1921) and Snelson (1948) (Snelson, 1965)."

> * [ ]  Please rephrase "Skelton defines “Class-1” tensegrity system where no compressive members touch each other, and a “Class-k” system, where k number of compressive members are connected at a node Skelton & Oliveira, 2009).", I propose (unless I misunderstood): "Skelton defines a tensegrity structure as a“Class-1” system if none of the compressive members are connected, if on the other hand k compressive members are connected at a node, this is referred to as a “Class-k” system (Skelton & Oliveira, 2009)."

The paper now reads:  "Skelton defines a tensegrity structure as a“Class-1” system if none of the compressive members are connected, if, on the other hand, $k$ compressive members are connected at a node, this is referred to as a “Class-$k$” system (Skelton & Oliveira, 2009)."

> * [ ]  Please change `... allowing to change the stiffness without changing the shape.` to `... allowing one to change the stiffness without changing the shape.`

We have made the recommended change.

> * [ ]  Please change `These properties of tensegrity structure have led the researchers to use tensegrity concepts in various applications from...` to :  `These properties of tensegrity structures have led researchers to use tensegrity concepts in various applications from...`

We have made the recommended change.
The sentence now reads: `These properties of tensegrity structures have led researchers to use tensegrity concepts in various applications from...`

> * [ ]  Please rephrase: `First, the static analysis provides the minimum mass of the tensegrity structure by optimizing for the tensile force in the strings and compressive force in the bars for no external force (self-equilibrium state) and in the presence of a given external force. The optimization problem is written as a Linear Programming to solve for the minimum mass required under yielding constraints.`

>   e.g. to `Firstly, static analysis provides the minimum mass of the tensegrity structure by optimizing tensile forces in the strings and compressive forces in the bars in the absence of external forces (self-equilibrium state), and in the presence of a given external forces.`

Agreed. The sentence now reads:
"Firstly, static analysis provides the minimum mass of the tensegrity structure by optimizing tensile forces in the strings and compressive forces in the bars in the absence of external forces (self-equilibrium state), and in the presence of given external forces."

> * [ ]  Please rephrase: `The optimization problem is written as a Linear Programming to solve for the minimum mass required under yielding constraints.` in particular `as a linear programming` sounds off. Should this be something like the following: `In order to solve for the minimum mass required under yielding constraints, MOTES formulates the optimization problem as a "Linear Programming" problem.`

Agreed. The sentence now reads:
"In order to solve for the minimum mass required under yielding constraints, MOTES formulates the optimization problem as a "Linear Programming" problem."

> * [ ]  Use `Secondly` instead of `Second` in "Second, the dynamic analysis"

Done. Added "Secondly".

> * [ ]  Can you rephrase this `... and strings to show elastic behavior (Hookean)` (if you agree) to `...and strings to exhibit linear elastic behaviour`.

Agreed. The sentence now reads:
`...and strings to exhibit linear elastic behavior`.

> * [ ]  Should `...the dynamics response...` be `...the dynamic response...`?

Yes, corrected everywhere in the paper.

> * [ ]  Instead of `The other disadvantage... `, i would use `Another disadvantage...`

Thanks, updated.

> * [ ]  Explain the acronym DOF

The sentence now reads:
"The approach uses 6 degrees of freedom (DOF)..."

> * [ ]  Change this `The use of non-minimal system results in...` to something like `The use of the non-minimal system results in...` or `The use of non-minimal systems results in...`

The sentence now reads:
"The use of the non-minimal system results..."

> * [ ]  `...the approximated errors at each...` should probably be `...the approximation errors at each...`?

The word "approximation" is the correct fit. Thanks!

> * [ ]  Please rephrase `...Tensegrity robotic arm, Tensegrity antenna, Tensegrity lander, and Space Habitat, where we integrate structure and control design to get the required performance.` to `...tensegrity robotic arms, tensegrity antennae, tensegrity landers, and space habitats, where we integrate structure and control design to get the required performance.`

All the capital letters have been changed to smaller ones. The sentence now reads:
"...tensegrity robotic arm, tensegrity antenna, tensegrity lander, and space habitat, where we integrate structure and control design to get the required performance."

> * [ ]  When referring to class material at the end of the paper, can you add the university and discipline?

The university and discipline has been added. The sentence now reads:
"The software was also used as a part of the class curriculum for the course '\textit{AERO-489/689 Design Elective: Advanced Statics and Dynamics of Flexible Structures: Tensegrity Systems}' which was offered in spring 2019 as an undergraduate-graduate class in the department of aerospace engineering at Texas A&M University, College Station."
