### 2019-10-22 #1613 @Kevin-Mattheus-Moerman
@ramaniitrgoyal92 thanks for making those changes. At this point can you do the following:

- [ ] Can you archive a copy of the software on Zenodo and list the DOI of the archived version here? Can you make sure that the author list and the title of the archived version matches your paper?
- [ ] Can you specify what latest version tag of the reviewed and archived software is? Are we still at `v2.1` or has the version moved on?

Thanks
