### 2019-10-22 #1613 @ramaniitrgoyal92
@Kevin-Mattheus-Moerman
Yes, I have uploaded the software to Zenodo, the link is: https://zenodo.org/record/3516295#.Xa8wIOhKhPY
Yes, we have made sure the author list and the title of the archived version match our paper.
The version is still v2.1.

Thank you!
