### 2019-10-23 #1613 @Kevin-Mattheus-Moerman
@ramaniitrgoyal92 it looks like you used the "reserved" JOSS paper DOI for your Zenodo archive. If possible can you amend the Zenodo archive to carry its own Zenodo DOI? (You may have to remove this archived version and create a new one with its own Zenodo DOI). Once you've done this can you report back here with the Zenodo DOI? Thanks.
