### 2019-10-23 #1613 @Kevin-Mattheus-Moerman
@openjournals/joss-eics this author changes the names of the project and the paper title from: `TEAM: Tensegrity Engineering Analysis Master` to `MOTES: Modeling of Tensegrity Structures`.

Does this require any special treatment? If this is dealt with this is paper is good to be accepted.
