### 2019-10-23 #1613 @kthyng
@arfon ok thanks.

@ramaniitrgoyal92 Can you go through your paper in detail and fix the citations? In a number of places, the spacing isn't correct around the references inline, and sometimes you use the parenthetical reference instead of inline, or vice versa. You can find examples for this here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography.

Also in your first paragraph it looks like you reference Ioganson (1921) and Snelson (1948) but by name instead of by citation, and then reference @Snelson_1965. It would make more sense, if I am reading this correctly, to cite inline Ioganson and Snelson 1948 directly and not reference Snelson 1965.

After you give your paper a careful review, ping me and I'll check it again, thanks.
