+++ JOSS_Paper/paper.bib
@@ -1,6 +1,6 @@
 @inproceedings{Goyal_2017_Habitat,
   title={Design and Control of Growth Adaptable Artificial Gravity Space Habitat},
-  author={Goyal, Raman and Bryant, Tyler and Majji, Manoranjan and Skelton, Robert E and Longman, Anthony},
+  author={Goyal, R. and Bryant, T. and Majji, M. and Skelton, R. and Longman, A.},
   booktitle={AIAA SPACE and Astronautics Forum and Exposition},
   pages={5141},
   year={2017},
@@ -9,7 +9,7 @@
 
 @inproceedings{Chen_2018_Habitat,
   title={Energy Analysis of Growth Adaptable Artificial Gravity Space Habitat},
-  author={Chen, Muhao and Shen, Yuling and Goyal, Raman  and Majji, Manoranjan and Skelton, Robert E.},
+  author={Chen, Muhao and Shen, Yuling and Goyal, Raman and Majji, Manoranjan and Skelton, Robert E.},
   booktitle={AIAA SPACE and Astronautics Forum and Exposition},
   pages={5109},
   year={2018},
@@ -27,7 +27,7 @@ year={2018}
 }
 
 @article{Karnan_Goyal_2017_IROS, 
-author={H. Karnan and R. Goyal and M. Majji and R. E. Skelton and P. Singla}, 
+author={Karnan, H. and Goyal, R. and Majji, M. and Skelton, R. E. and Singla, P.}, 
 journal={2017 IEEE/RSJ-IROS}, 
 title={Visual feedback control of tensegrity robotic systems}, 
 year={2017}, 
@@ -39,7 +39,7 @@ month={Sept},
 
 @article{Goyal_2019_Buckling,
   title={Analytical Study of Tensegrity Lattices for Mass-Efficient Mechanical Energy Absorption},
-  author={Goyal, R. and Hernandez,Edwin A. Peraza and Skelton, R. E.},
+  author={Goyal, R. and Hernandez, E. A. P. and Skelton, R. E.},
   journal={International Journal of Space Structures},
   pages={1-19},
   DOI = {10.1177/0956059919845330},
@@ -47,7 +47,7 @@ month={Sept},
 }
 
 @INPROCEEDINGS{Sabelhaus, 
-author={A. P. {Sabelhaus} and A. K. {Akella} and Z. A. {Ahmad} and V. {SunSpiral}}, 
+author={Sabelhaus, A. P. and Akella, A. K. and Ahmad, Z. A. and SunSpiral, V.}, 
 booktitle={2017 American Control Conference (ACC)}, 
 title={Model-Predictive Control of a flexible spine robot}, 
 year={2017}, 
@@ -136,7 +136,7 @@ month={May},}
 
 @article{Stedy_2019,
   title={STEDY: Software for TEnsegrity DYnamics},
-  author={Vaishnav Tadiparthi, Shao-Chen Hsu, and Raktim Bhattacharya},
+  author={Tadiparthi, V. and Hsu, S.C. and Bhattacharya, R.},
   journal={Journal of Open Source Software},
   year={2019},
   volume={4(33)},
