+++ JOSS_Paper/paper.bib
@@ -1,4 +1,4 @@
-@inproceedings{Goyal_2017_Habitat,
+@article{Goyal_2017,
   title={Design and Control of Growth Adaptable Artificial Gravity Space Habitat},
   author={Goyal, R. and Bryant, T. and Majji, M. and Skelton, R. and Longman, A.},
   booktitle={AIAA SPACE and Astronautics Forum and Exposition},
@@ -7,7 +7,7 @@
   doi ={10.2514/6.2017-5141}, 
 }
 
-@inproceedings{Chen_2018_Habitat,
+@article{Chen_2018_Habitat,
   title={Energy Analysis of Growth Adaptable Artificial Gravity Space Habitat},
   author={Chen, Muhao and Shen, Yuling and Goyal, Raman and Majji, Manoranjan and Skelton, Robert E.},
   booktitle={AIAA SPACE and Astronautics Forum and Exposition},
@@ -16,7 +16,7 @@
   doi ={10.2514/6.2018-5109},
 }
 
-@inproceedings{Goyal_2018_ASCE,
+@article{Goyal_2018_ASCE,
 title={Dynamics of Class 1 Tensegrity Systems Including Cable Mass},
 author={Goyal, R. and Skelton, R. E.},
 booktitle={16th Biennial ASCE Conference - Earth and Space 2018},
@@ -26,7 +26,7 @@ URL = {https://ascelibrary.org/doi/abs/10.1061/9780784481899.082},
 year={2018}
 }
 
-@article{Karnan_Goyal_2017_IROS, 
+@article{Karnan_2017, 
 author={Karnan, H. and Goyal, R. and Majji, M. and Skelton, R. E. and Singla, P.}, 
 journal={2017 IEEE/RSJ-IROS}, 
 title={Visual feedback control of tensegrity robotic systems}, 
@@ -46,7 +46,7 @@ month={Sept},
   year={2019}
 }
 
-@INPROCEEDINGS{Sabelhaus, 
+@INPROCEEDINGS{Sabelhaus_2017, 
 author={Sabelhaus, A. P. and Akella, A. K. and Ahmad, Z. A. and SunSpiral, V.}, 
 booktitle={2017 American Control Conference (ACC)}, 
 title={Model-Predictive Control of a flexible spine robot}, 
@@ -95,7 +95,7 @@ month={May},}
   organization={International Society for Optics and Photonics}
 }
 
-@article{Carpentieri_2015_Bridge,
+@article{Carpentieri_2015,
   title={Minimum mass and optimal complexity of planar tensegrity bridges},
   author={Carpentieri, G. and Skelton, R. E. and Fraternali, F.},
   journal={International Journal of Space Structures},
@@ -134,7 +134,7 @@ month={May},}
   doi={10.2514/6.2003-1978}
 }
 
-@article{Stedy_2019,
+@article{Tadiparthi_2019,
   title={STEDY: Software for TEnsegrity DYnamics},
   author={Tadiparthi, V. and Hsu, S.C. and Bhattacharya, R.},
   journal={Journal of Open Source Software},
