### 2019-10-24 bb80b8a
Update paper.md
