+++ JOSS_Paper/paper.bib
@@ -143,3 +143,13 @@ month={May},}
   pages={1042},
   doi={10.21105/joss.01042}
 }
+
+@article{Sultan_2009,
+  title={Tensegrity Structures: Sixty Years of Art, Science, and Engineering},
+  author={Sultan, Cornel},
+  journal={Advances in Applied Mechanics},
+  year={2009},
+  volume={43)},
+  pages={69-145},
+  doi={10.1016/S0065-2156(09)43002-3}
+}
