### 2019-10-24 ac41f32
Update paper.md
