+++ JOSS_Paper/paper.md
@@ -36,7 +36,8 @@ bibliography: paper.bib
 
 # Tensegrity Structures
 
-Tensegrity system dynamics is a subset of the class of multi-body dynamics which includes cylindrical rigid bodies (bars) and elastic members (strings) arranged in a stabilizable topology [@Skelton_2009_Tensegrity_Book]. The term tensegrity was first coined by Buckminster @Fuller_1959, and is a portmanteau of "tensional integrity". Tensegrity structures also appeared in earlier artworks by Ioganson and @Snelson_1965. Tensegrity structures are created by methodically arranging tensile members (strings) and compressive members (bars) to form a stable system. @Skelton_2009_Tensegrity_Book define a tensegrity structure as a “class-1” tensegrity system if none of the compressive members are connected, if, on the other hand, $k$ compressive members are connected at a node, this is referred to as a “class-$k$” tensegrity system. The multiple compressive members are connected through ball joints, causing all the members in a tensegrity structure to be axially loaded, i.e., no moment is present on any individual member. Tensegrity structures can also be prestressed to have uni-directional loading for all the individual members giving the freedom to design tension and compression members separately. This freedom provides good structural efficiency (high strength-to-mass ratio) to tensegrity structures. A particular topology of tensegrity structure can provide various self-equilibrium solutions corresponding to different values of prestress in the structure. The various stable prestress values provide a domain set to minimize the mass of the structure. This minimization has proved tensegrity to be an optimal mass solution for various loading conditions [@Skelton_2009_Tensegrity_Book]. Moreover, different prestress values correspond to different stiffness, allowing one to change the stiffness without changing the shape. These properties of tensegrity structure have led researchers to use tensegrity concepts in various applications from civil engineering bridges [@Carpentieri_2015] to soft-robotics [@Sabelhaus_2017; @Karnan_2017] to various space applications; landers [@Vytas_2013; @Goyal_2019_Buckling], deployable structures [@Tibert_Pellegrino_2003_Mast], and, space habitat designs [@Goyal_2017].
+Tensegrity system dynamics is a subset of the class of multi-body dynamics which includes cylindrical rigid bodies (bars) and elastic members (strings) arranged in a stabilizable topology [@Skelton_2009_Tensegrity_Book]. Tensegrity structures appeared in earlier artworks by Karl Ioganson in 1921 and Kenneth Snelson in 1948 [@Sultan_2009]. The term tensegrity was first coined by Buckminster @Fuller_1959, and is a portmanteau of "tensional integrity" and @Snelson_1965 filed the first patents on this topic.
[…]
