### 2019-10-24 #1613 @kthyng
This looks reasonable to me. Sorry I originally misunderstood your intent, but I like it better now anyway :)
