### 2019-10-24 #1613 @kthyng
I still don't see the DOI link working, but I need to move on for now. I will check back later.
