### 2019-10-25 #1613 @arfon
@kthyng - things look to be resolving OK now.
